Fix bug

+++ src/main/java/view/rmi/RMIViewTimer.java
@@ -33,8 +33,8 @@ public class RMIViewTimer implements Runnable {
 		this.timer2.cancel();
 		this.timer1 = new Timer();
 		this.timer2 = new Timer();
-		this.timer1.schedule(new RMITimerAdvice(this.client, this.server, this.player), 10*1000);
-		this.timer2.schedule(new RMIDisconnect(this.client, this.server, this.player), 20*1000);
+		this.timer1.schedule(new RMITimerAdvice(this.client, this.server, this.player), (long)10*1000);
+		this.timer2.schedule(new RMIDisconnect(this.client, this.server, this.player), (long)20*1000);
 		System.out.println("Timer reset");
 	}
 	
@@ -48,8 +48,8 @@ public class RMIViewTimer implements Runnable {
 	public void start() throws RemoteException{
 		this.timer1 = new Timer();
 		this.timer2 = new Timer();
-		this.timer1.schedule(new RMITimerAdvice(this.client, this.server, this.player), 10*1000);
-		this.timer2.schedule(new RMIDisconnect(this.client, this.server, this.player), 20*1000);
+		this.timer1.schedule(new RMITimerAdvice(this.client, this.server, this.player), (long)10*1000);
+		this.timer2.schedule(new RMIDisconnect(this.client, this.server, this.player), (long)20*1000);
 		this.active = true;
 		System.out.println("Timer start");
 	}
